requirements.txt
beautifulsoup4==4.9.3
cachetools==4.2.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.6
filelock==3.8.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==2.10
oauthlib==3.1.1
platformdirs==2.5.2
protobuf==3.19.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
PyQt6==6.2.1
PyQt6-Qt6==6.2.1
PyQt6-sip==13.1.0
python-dotenv==0.19.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.2.1
uritemplate==4.1.1
urllib3==1.26.4
virtualenv==20.16.5
